fix: prevent float16 overflow in router/pooler during MoC training
After the first optimizer step, the router and pooler (converted to
float16 by moc.half()) overflow when computing q and r, producing
NaN L_lb for every subsequent batch.

Fix 1 (train_moc.py): after moc.half(), immediately call
moc.router.float() and moc.pooler.float() to keep these two small
modules in float32.  The expert modules (E2, E3, E4) stay float16
for memory efficiency.

Fix 2 (connector/moc.py): split the single _dtype variable into
_router_dtype (from router.W1.weight, float32 in training) and
_expert_dtype (from experts[1], float16 in training).  question_embeddings
is cast to _router_dtype before the pooler; Z_V and q are cast to
_expert_dtype before the expert forward.  This cleanly separates
numerical-stability concerns (router/pooler) from memory concerns
(expert projection matrices).

+++ train_moc.py
@@ -407,7 +407,9 @@ def train(args):
     # ---- 3. Apply QLoRA ------------------------------------------------------
     print("Applying QLoRA …")
     model = prepare_model_for_kbit_training(model)
-    moc.half()   # re-cast after kbit prep upcasted fp16 params to fp32
+    moc.half()         # convert all MoC params to float16 (matches LLM compute dtype)
+    moc.router.float() # but keep router in float32 — float16 overflows after first step
+    moc.pooler.float() # keep pooler in float32 for the same reason
 
     lora_cfg = LoraConfig(
         r=16, lora_alpha=32, lora_dropout=0.05,
